reorganize dervied market data to get ready for mma markets and other market factories
